+++ contracts/universalSchemes/SchemeRegistrar.sol
@@ -11,7 +11,7 @@ import "../votingMachines/VotingMachineCallbacks.sol";
  * @dev The SchemeRegistrar is used for registering and unregistering schemes at organizations
  */
 
-contract SchemeRegistrar is UniversalScheme,VotingMachineCallbacks,ProposalExecuteInterface {
+contract SchemeRegistrar is UniversalScheme, VotingMachineCallbacks, ProposalExecuteInterface {
     event NewSchemeProposal(
         address indexed _avatar,
         bytes32 indexed _proposalId,
@@ -20,12 +20,14 @@ contract SchemeRegistrar is UniversalScheme,VotingMachineCallbacks,ProposalExecu
         bytes32 _parametersHash,
         bytes4 _permissions
     );
+
     event RemoveSchemeProposal(address indexed _avatar,
         bytes32 indexed _proposalId,
         address indexed _intVoteInterface,
         address _scheme
     );
-    event ProposalExecuted(address indexed _avatar, bytes32 indexed _proposalId,int _param);
+
+    event ProposalExecuted(address indexed _avatar, bytes32 indexed _proposalId, int256 _param);
     event ProposalDeleted(address indexed _avatar, bytes32 indexed _proposalId);
 
     // a SchemeProposal is a  proposal to add or remove a scheme to/from the an organization
@@ -45,6 +47,7 @@ contract SchemeRegistrar is UniversalScheme,VotingMachineCallbacks,ProposalExecu
         bytes32 voteRemoveParams;
         IntVoteInterface intVote;
     }
+
     mapping(bytes32=>Parameters) public parameters;
 
     /**
@@ -52,12 +55,12 @@ contract SchemeRegistrar is UniversalScheme,VotingMachineCallbacks,ProposalExecu
     * @param _proposalId the ID of the voting in the voting machine
     * @param _param a parameter of the voting result, 1 yes and 2 is no.
     */
-    function executeProposal(bytes32 _proposalId,int _param) external onlyVotingMachine(_proposalId) returns(bool) {
+    function executeProposal(bytes32 _proposalId, int256 _param) external onlyVotingMachine(_proposalId) returns(bool) {
         Avatar avatar = proposalsInfo[_proposalId].avatar;
         SchemeProposal memory proposal = organizationsProposals[address(avatar)][_proposalId];
         require(proposal.scheme != address(0));
         delete organizationsProposals[address(avatar)][_proposalId];
-        emit ProposalDeleted(address(avatar),_proposalId);
+        emit ProposalDeleted(address(avatar), _proposalId);
         if (_param == 1) {
 
           // Define controller and get the params:
@@ -65,14 +68,19 @@ contract SchemeRegistrar is UniversalScheme,VotingMachineCallbacks,ProposalExecu
 
           // Add a scheme:
             if (proposal.proposalType == 1) {
-                require(controller.registerScheme(proposal.scheme, proposal.parametersHash, proposal.permissions,address(avatar)));
+                require(controller.registerScheme(
+                        proposal.scheme,
+                        proposal.parametersHash,
+                        proposal.permissions,
+                        address(avatar))
+                );
             }
           // Remove a scheme:
-            if ( proposal.proposalType == 2 ) {
-                require(controller.unregisterScheme(proposal.scheme,address(avatar)));
+            if (proposal.proposalType == 2) {
+                require(controller.unregisterScheme(proposal.scheme, address(avatar)));
             }
-          }
-        emit ProposalExecuted(address(avatar), _proposalId,_param);
+        }
+        emit ProposalExecuted(address(avatar), _proposalId, _param);
         return true;
     }
 
@@ -144,10 +152,11 @@ contract SchemeRegistrar is UniversalScheme,VotingMachineCallbacks,ProposalExecu
             _permissions
         );
         organizationsProposals[address(_avatar)][proposalId] = proposal;
-        proposalsInfo[proposalId] = ProposalInfo(
-            {blockNumber:block.number,
+        proposalsInfo[proposalId] = ProposalInfo({
+            blockNumber:block.number,
             avatar:_avatar,
-            votingMachine:address(controllerParams.intVote)});
+            votingMachine:address(controllerParams.intVote)
+        });
         return proposalId;
     }
 
@@ -166,15 +175,16 @@ contract SchemeRegistrar is UniversalScheme,VotingMachineCallbacks,ProposalExecu
         Parameters memory params = parameters[paramsHash];
 
         IntVoteInterface intVote = params.intVote;
-        bytes32 proposalId = intVote.propose(2, params.voteRemoveParams,msg.sender,address(_avatar));
+        bytes32 proposalId = intVote.propose(2, params.voteRemoveParams, msg.sender, address(_avatar));
 
         organizationsProposals[address(_avatar)][proposalId].proposalType = 2;
         organizationsProposals[address(_avatar)][proposalId].scheme = _scheme;
         emit RemoveSchemeProposal(address(_avatar), proposalId, address(intVote), _scheme);
-        proposalsInfo[proposalId] = ProposalInfo(
-            {blockNumber:block.number,
+        proposalsInfo[proposalId] = ProposalInfo({
+            blockNumber:block.number,
             avatar:_avatar,
-            votingMachine:address(params.intVote)});
+            votingMachine:address(params.intVote)
+        });
         return proposalId;
     }
 }
